test(providers): target volar.ts mutation gaps + split test helpers
Split tests/helpers.ts into three cohesive modules:
- tests/helpers.ts — fixture I/O only (copyFixture, cleanup, readFile,
  fileExists, PROJECT_ROOT)
- tests/process-helpers.ts — CLI spawn and daemon helpers
  (spawnAndWaitForReady, waitForDaemon, killDaemon, callDaemonSocket,
  runCliCommand)
- tests/mcp-helpers.ts — MCP client (McpTestClient, parseMcpResult,
  useMcpContext)

Add two VolarProvider tests targeting NoCoverage mutation gaps in volar.ts:
- getDefinitionAtPosition on a template-only .vue file (no <script> block)
  exercises the toVirtualLocation !serviceScript fallback branch using a
  programmatic fixture instead of modifying the shared vue-project fixture
- getRenameLocations on a blank-line offset in a .ts file exercises the
  rawLocs.length === 0 null-return guard; getRenameLocations requires .ts
  paths — the Volar proxy LS only accepts .vue.ts virtual paths, not .vue

322/322 tests passing. Re-run pnpm test:mutate for an updated volar.ts score.
