+++ tests/mcp/get-definition.test.ts
@@ -1,6 +1,6 @@
 import * as path from "node:path";
 import { describe, expect, it } from "vitest";
-import { parseMcpResult, useMcpContext } from "../helpers";
+import { parseMcpResult, useMcpContext } from "../mcp-helpers.js";
 
 describe("MCP transport — getDefinition tool", () => {
   const { setup } = useMcpContext();
